+++ src/lib/conversions.ts
@@ -14,7 +14,7 @@ const colorTemperature_out: PropertyTransformKernel = (value) => {
 	const [min, max] = colorTemperatureRange;
 	// extrapolate 0-100% to [min..max]
 	value = clamp(value, 0, 100);
-	return roundTo(min + value / 100 * (max - min), 0);
+	return roundTo(min + (value / 100) * (max - min), 0);
 };
 // interpolate from [250..454] to [0..100%]
 const colorTemperature_in: PropertyTransformKernel = (value) => {
@@ -104,7 +104,7 @@ const colorTemperature_in: PropertyTransformKernel = (value) => {
 
 function rgbToHSV(r: number, g: number, b: number) {
 	// transform [0..255] => [0..1]
-	[r, g, b] = [r, g, b].map(c => c / 255);
+	[r, g, b] = [r, g, b].map((c) => c / 255);
 	const max = Math.max(r, g, b);
 	const min = Math.min(r, g, b);
 	let h: number;
@@ -127,7 +127,7 @@ function rgbToHSV(r: number, g: number, b: number) {
 	} else {
 		s = roundTo((max - min) / max, 3);
 	}
-	return {h, s, v};
+	return { h, s, v };
 }
 
 function rgbFromHSV(h: number, s: number, v: number) {
@@ -138,34 +138,46 @@ function rgbFromHSV(h: number, s: number, v: number) {
 		r = g = b = v;
 	} else {
 		const hi = Math.floor(h / 60);
-		const f = (h / 60 - hi);
+		const f = h / 60 - hi;
 		const p = v * (1 - s);
 		const q = v * (1 - s * f);
 		const t = v * (1 - s * (1 - f));
 		switch (hi) {
 			case 0:
-			case 6: [r, g, b] = [v, t, p]; break;
-			case 1: [r, g, b] = [q, v, p]; break;
-			case 2: [r, g, b] = [p, v, t]; break;
-			case 3: [r, g, b] = [p, q, v]; break;
-			case 4: [r, g, b] = [t, p, v]; break;
-			case 5: [r, g, b] = [v, p, q]; break;
+			case 6:
+				[r, g, b] = [v, t, p];
+				break;
+			case 1:
+				[r, g, b] = [q, v, p];
+				break;
+			case 2:
+				[r, g, b] = [p, v, t];
+				break;
+			case 3:
+				[r, g, b] = [p, q, v];
+				break;
+			case 4:
+				[r, g, b] = [t, p, v];
+				break;
+			case 5:
+				[r, g, b] = [v, p, q];
+				break;
 		}
 	}
 	// transform back to [0..255] - r, g and b are defined always
-	[r, g, b] = [r!, g!, b!].map(c => Math.round(clamp(c, 0, 1) * 255));
-	return {r, g, b};
+	[r, g, b] = [r!, g!, b!].map((c) => Math.round(clamp(c, 0, 1) * 255));
+	return { r, g, b };
 }
 
 function rgbToString(r: number, g: number, b: number): string {
-	return [r, g, b].map(c => padStart(c.toString(16), 2, "0")).join("");
+	return [r, g, b].map((c) => padStart(c.toString(16), 2, "0")).join("");
 }
 
 function rgbFromString(rgb: string) {
 	const r = parseInt(rgb.substr(0, 2), 16);
 	const g = parseInt(rgb.substr(2, 2), 16);
 	const b = parseInt(rgb.substr(4, 2), 16);
-	return {r, g, b};
+	return { r, g, b };
 }
 
 // ===========================
@@ -175,7 +187,7 @@ const hue_out: PropertyTransformKernel = (value, light?: Light) => {
 	if (light != null && light.spectrum !== "rgb") return null; // hue is not supported
 
 	value = clamp(value, 0, 360);
-	return roundTo(value / 360 * MAX_COLOR, 0);
+	return roundTo((value / 360) * MAX_COLOR, 0);
 };
 // interpolate hue from [0..COLOR_MAX] to [0..360]
 const hue_in: PropertyTransformKernel = (value /*, light: Light*/) => {
@@ -188,7 +200,7 @@ const saturation_out: PropertyTransformKernel = (value, light?: Light) => {
 	if (light != null && light.spectrum !== "rgb") return null; // hue is not supported
 
 	value = clamp(value, 0, 100);
-	return roundTo(value / 100 * MAX_COLOR, 0);
+	return roundTo((value / 100) * MAX_COLOR, 0);
 };
 // interpolate saturation from [0..COLOR_MAX] to [0..100%]
 const saturation_in: PropertyTransformKernel = (value /*, light: Light*/) => {
@@ -200,9 +212,9 @@ const saturation_in: PropertyTransformKernel = (value /*, light: Light*/) => {
 // TRANSITION TIME conversions
 
 // the sent value is in 10ths of seconds, we're working with seconds
-const transitionTime_out: PropertyTransformKernel = val => val && val * 10; // "val && " avoids sending `null` if val is null for some reason
+const transitionTime_out: PropertyTransformKernel = (val) => val && val * 10; // "val && " avoids sending `null` if val is null for some reason
 // the sent value is in 10ths of seconds, we're working with seconds
-const transitionTime_in: PropertyTransformKernel = val => val / 10;
+const transitionTime_in: PropertyTransformKernel = (val) => val / 10;
 
 // ===========================
 // BRIGHTNESS conversions
@@ -210,13 +222,13 @@ const transitionTime_in: PropertyTransformKernel = val => val / 10;
 // interpolate from [0..100%] to [0..254]
 const brightness_out: PropertyTransformKernel = (value) => {
 	value = clamp(value, 0, 100);
-	return roundTo(value / 100 * 254, 0);
+	return roundTo((value / 100) * 254, 0);
 };
 // interpolate from [0..254] to [0..100%]
 const brightness_in: PropertyTransformKernel = (value) => {
 	value = clamp(value, 0, 254);
 	if (value === 0) return 0;
-	value = value / 254 * 100;
+	value = (value / 254) * 100;
 	return roundTo(value, 1);
 };
 
@@ -225,9 +237,31 @@ const brightness_in: PropertyTransformKernel = (value) => {
 
 // The gateway expects 0 to be open and 100 to be closed
 // we do the opposite
-const position_out: PropertyTransformKernel = val => 100 - val;
+const position_out: PropertyTransformKernel = (val) => 100 - val;
 // the sent value is in 10ths of seconds, we're working with seconds
 const position_in: PropertyTransformKernel = position_out;
+
+// ===========================
+// AIR PURIFIER conversions
+
+// The fan speed only accepts certain values
+const fan_speed_out: PropertyTransformKernel = (val: number) => {
+	if (val === 0) return 0;
+	if (val < 10) return 10;
+	if (val > 50) return 50;
+	// Round to the nearest multiple of 5
+	return Math.round(val / 5) * 5;
+};
+
+// ===========================
+// PRIMITIVE conversions
+
+// Some values are boolean but have an inverted meaning
+const boolean_inverted_out: PropertyTransformKernel = (bool: boolean) =>
+	bool ? 0 : 1;
+// the sent value is in 10ths of seconds, we're working with seconds
+const boolean_inverted_in: PropertyTransformKernel = (raw: any) =>
+	!(raw === 1 || raw === "true" || raw === "on" || raw === true);
 
 export const serializers = {
 	transitionTime: transitionTime_out,
@@ -236,6 +270,8 @@ export const serializers = {
 	brightness: brightness_out,
 	colorTemperature: colorTemperature_out,
 	position: position_out,
+	booleanInverted: boolean_inverted_out,
+	fanSpeed: fan_speed_out,
 };
 
 export const deserializers = {
@@ -245,6 +281,7 @@ export const deserializers = {
 	brightness: brightness_in,
 	colorTemperature: colorTemperature_in,
 	position: position_in,
+	booleanInverted: boolean_inverted_in,
 };
 
 export const conversions = {