+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 	Placeholder for next release:
 	## __WORK IN PROGRESS__
 -->
+## __WORK IN PROGRESS__
+* (AlCalzone) Added support for STARKVIND air purifier
+
 ## 3.0.1 (2021-06-30)
 * (AlCalzone) Workaround for a bug in IKEA gateway firmware `v1.15.x`